Extended the CLUSTERIP module to use it on a IPSec gateway cluster
The current CLUSTERIP implementation works fine for INPUT/OUTPUT traffic,
but is unable to process traffic as an IPsec gateway. These changes
add support to the CLUSTERIP module to use it for traffic processing
on an IPsec gateway. Multicast packets are tagged for forwarding and
forwarded traffic is dropped just before XFRM processing. XFRM sequence
numbers are kept in sync, allowing a cluster to do load sharing of IPsec
traffic.