netfilter: Add hooks before XFRM input and output processing
To allow netfilter hooks to modify, inspect or drop packets before
XFRM processing, two hooks are added to the netfilter framework.
The first is invoked just before XFRM input processing, the other
before XFRM output processing. This allows a hook to drop the packet,
but mangle the associated xfrm state (sequence numbers, for example).

NF_INET_NUMHOOKS does not include the new hooks. This does not allow the use
of the new hooks from userspace, but keeps the ABI as it is.

Signed-off-by: Martin Willi <martin@strongswan.org>