netfilter: ipt_CLUSTERIP: Add CLUSTERIP support for IPsec packets
The current CLUSTERIP implementation is limited to INPUT/OUTPUT traffic and
can't segmentate traffic as an IPsec gateway. These changes add support to the
CLUSTERIP module for traffic subject to XFRM processing. Multicast packets are
tagged for forwarding and forwarded traffic is dropped just before XFRM
processing. IPsec sequence numbers are kept in sync, allowing a node to
transparently take over the security associations of a failing node.

To avoid packet reordering, packets for a single security association get
processed on the same node. This is realized by including the SPI and
destination address in the ClusterIP hash, but not the IPsec sequence number.

Signed-off-by: Martin Willi <martin@strongswan.org>